+++ src/zephyr/client.ts
@@ -7,10 +7,12 @@ import type {
 import { ApiClient } from "./common/api-client";
 
 import { GetEnvironments } from "./tool/environment/get-environments";
+import { CreateFolder } from "./tool/folder/create-folder";
 import { GetPriorities } from "./tool/priority/get-priorities";
 import { GetProject } from "./tool/project/get-project";
 import { GetProjects } from "./tool/project/get-projects";
 import { GetStatuses } from "./tool/status/get-statuses";
+import { CreateTestCaseIssueLink } from "./tool/test-case/create-issue-link";
 import { CreateTestCase } from "./tool/test-case/create-test-case";
 import { CreateTestSteps } from "./tool/test-case/create-test-steps.ts";
 import { CreateTestCaseWebLink } from "./tool/test-case/create-web-link.ts";
@@ -87,7 +89,12 @@ export class ZephyrClient implements Client {
       new UpdateTestCycle(this),
       new CreateTestExecution(this),
       new CreateTestCaseWebLink(this),
+<<<<<<< HEAD
       new CreateTestSteps(this),
+=======
+      new CreateTestCaseIssueLink(this),
+      new CreateFolder(this),
+>>>>>>> main
     ];
 
     for (const tool of tools) {